==> 01_train_validation_split.ipynb <==
import os
import pandas as pd
import numpy as np
import cv2
import random
import shutil
import matplotlib.pyplot as plt
import matplotlib.image as mpimg

#creating validation set

os.makedirs('validation')

subfolders = [f.path for f in os.scandir('train') if f.is_dir() ] 

subfolders

os.listdir('train')

for i in os.listdir('train'):
    os.makedirs('validation/'+i)

file_list = [list(sorted(os.listdir('train/' + x))) for x in sorted(os.listdir('train'))]
np.random.seed(10)
fold_list = [[np.random.choice(range(5)) for _ in x] for x in file_list]
this_fold=1
train_files = [[y for j,y in enumerate(x) if fold_list[i][j] != this_fold] for i,x in enumerate(file_list)]
val_files   = [[y for j,y in enumerate(x) if fold_list[i][j] == this_fold] for i,x in enumerate(file_list)]

train_files

val_files

for i,x in enumerate(sorted(os.listdir('train'))):
    for y in val_files[i]:
        os.rename('train/' + x + '/' + y, 'validation/' + x + '/' + y)

#center crop for validation set

for i in os.listdir('validation/'):
    for j in os.listdir('validation/' + i):
        image = mpimg.imread('validation/' + i + '/' + j)
        h,w,d = image.shape
        image_crop = image[h//2-256:h//2+256,w//2-256:w//2+256,:]
        image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
        cv2.imwrite('validation/' + i + '/' + j, image_crop)

#random crop for training data

for i in os.listdir('train/'):
    print(i)
    for j in os.listdir('train/' + i):
        image = mpimg.imread('train/' + i + '/' + j)
        if(len(image.shape)<3):
            continue
        h,w,d = image.shape
        for k in range(100):
            rand_num_h = random.randint(0,h-512)
            rand_num_w = random.randint(0,w-512)
            image_crop = image[rand_num_h:rand_num_h+512,rand_num_w:rand_num_w+512,:]
            image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
            if(k==0):
                cv2.imwrite('train/' + i + '/' + j, image_crop)
            else:
                cv2.imwrite('train/' + i + '/' + str(k) + '_'+ j, image_crop)

#cropping test set into five parts for predictions 224x224 size

# move all files in test folder to test/temp/

for i in os.listdir('test'):
    for j in os.listdir('test/' + i):
        image = mpimg.imread('test/' + i + '/' + j)
        h,w,d = image.shape
        image_crop = image[0:224,0:224,:]
        image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
        cv2.imwrite('test/' + i + '/' + '1' + '_' + j, image_crop)
        image_crop = image[0:224,-224:,:]
        image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
        cv2.imwrite('test/' + i + '/' + '2' + '_' + j, image_crop)
        image_crop = image[-224:,-224:,:]
        image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
        cv2.imwrite('test/' + i + '/' + '3' + '_' + j, image_crop)
        image_crop = image[-224:,0:224,:]
        image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
        cv2.imwrite('test/' + i + '/' + '4' + '_' + j, image_crop)
        image_crop = image[h//2-112:h//2+112,w//2-112:w//2+112,:]
        image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
        cv2.imwrite('test/' + i + '/' + '5' + '_' + j, image_crop)

#removing original test images

for i in os.listdir('test'):
    for j in os.listdir('test/' + i):
        if(len(j.split('_'))<4):
            os.remove('test/' + i+'/'+j)

#for cropping validation and train data into 224x224 chunks

for i in os.listdir('validation'):
    for j in os.listdir('validation/' + i):
        image = mpimg.imread('validation/' + i + '/' + j)
        h,w,d = image.shape
        temp_sum_h = 0
        while(temp_sum_h + 224 < h):
            temp_sum_w = 0
            while(temp_sum_w + 224 < w):
                image_crop = image[temp_sum_h:temp_sum_h+224, temp_sum_w:temp_sum_w+224]
                image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
                if(temp_sum_h==0 and temp_sum_w==0):
                    cv2.imwrite('validation/' + i + '/' + j, image_crop)
                else:
                    cv2.imwrite('validation/' + i + '/' + str(temp_sum_h) + str(temp_sum_w)+ '_' + j, image_crop)
                temp_sum_w=temp_sum_w+224
            temp_sum_h=temp_sum_h + 224  


for i in os.listdir('train'):
    for j in os.listdir('train/' + i):
        image = mpimg.imread('train/' + i + '/' + j)
        if(len(image.shape)<3):
            print(1)
            continue
        h,w,d = image.shape
        temp_sum_h = 0
        while(temp_sum_h + 224 < h):
            temp_sum_w = 0
            while(temp_sum_w + 224 < w):
                image_crop = image[temp_sum_h:temp_sum_h+224, temp_sum_w:temp_sum_w+224]
                image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
                if(temp_sum_h==0and temp_sum_w==0):
                    cv2.imwrite('train/' + i + '/' + j, image_crop)
                else:
                    cv2.imwrite('train/' + i + '/' + str(temp_sum_h) +  str(temp_sum_w)+'_' + j, image_crop)
                temp_sum_w=temp_sum_w+224
            temp_sum_h=temp_sum_h + 224  

# Downloading external data

subfolders = [f.path for f in os.scandir('../flickr_images/') if f.is_dir() ] 
subfolders

from collections import defaultdict
goodimgdict = defaultdict(list)
with open('../flickr_images/good_jpgs','r') as goodimgfile:
    goodimg = [i.replace('\n','') for i in goodimgfile.readlines()]
    for i in goodimg:
        goodimgdict[i.split('/')[-2]].append(i.split('/')[-1])

for j in subfolders:
    with open (j+'/urls_final', "r") as myfile:
        data = [i.replace('\n','') for i in myfile.readlines()]
        for i in data:
            if(i.split('/')[-1] in goodimgdict[j.split('/')[-1]]):
                os.system("cd "+ j + "\n wget " + i)
            

with open ('../moto_x/urls.txt', "r") as myfile:
    data = [i.replace('\n','') for i in myfile.readlines()]
    for i in data:
        os.system("cd ../moto_x/"  + "\n wget " + i)
    

subfolders = [f.path for f in os.scandir('../val_images/') if f.is_dir() ] 
os.listdir(subfolders[0])

for j in subfolders:
    x = os.listdir(j)
    with open (j+'/'+x[0], "r") as myfile:
        data = [i.replace('\n','') for i in myfile.readlines()]
        for i in data:
            os.system("cd "+ j + "\n wget " + i)
            

#removing url files in external data folders

for i in os.listdir('flickr_images/'):
    os.remove('flickr_images/' + i + '/'+'urls')

for i in os.listdir('flickr_images/'):
    os.remove('flickr_images/' + i + '/'+'urls_final')


for i in os.listdir('val_images/'):
    for j in os.listdir('val_images/' + i):
        if(j.split('_')[0]=='urls'):
            os.remove('val_images/' + i+'/'+j)

#cropping external data in the same way for given original data

# remove good_jpgs file in each folder

for i in os.listdir('flickr_images/'):
    print(i)
    for j in os.listdir('flickr_images/' + i):
        image = mpimg.imread('flickr_images/' + i + '/' + j)
        if(len(image.shape)<3):
            continue
        h,w,d = image.shape
        for k in range(100):
            rand_num_h = random.randint(0,h-512)
            rand_num_w = random.randint(0,w-512)
            image_crop = image[rand_num_h:rand_num_h+512,rand_num_w:rand_num_w+512,:]
            image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
            if(k==0):
                cv2.imwrite('flickr_images/' + i + '/' + j, image_crop)
            else:
                cv2.imwrite('flickr_images/' + i + '/' + str(k) + '_'+ j, image_crop)

# rename the jpg.{0-9} images to jpg

for i in os.listdir('val_images/'):
    for j in os.listdir('val_images/' + i):
        image = mpimg.imread('val_images/' + i + '/' + j)
        h,w,d = image.shape
        image_crop = image[h//2-256:h//2+256,w//2-256:w//2+256,:]
        image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
        cv2.imwrite('val_images/' + i + '/' + j, image_crop)

for k in range(1):
    for j in os.listdir('moto_x/'):
        image = mpimg.imread('moto_x/' + j)
        h,w,d = image.shape
        for k in range(100):
            rand_num_h = random.randint(0,h-512)
            rand_num_w = random.randint(0,w-512)
            image_crop = image[rand_num_h:rand_num_h+512,rand_num_w:rand_num_w+512,:]
            image_crop = cv2.cvtColor(image_crop, cv2.COLOR_RGB2BGR)
            cv2.imwrite('moto_x/' + str(k) + '_' + j, image_crop)

from PIL import Image
for k in range(1):
    for j in os.listdir('moto_x/'):
        im = Image.open('moto_x/' + j)
        if(im.size[0]>513):
            os.remove('moto_x/' + j)

# copying external data into main train and validation folders

for i in os.listdir('val_images/iphone_4s'):
    shutil.copy('val_images/iphone_4s/'+i,'validation/iPhone-4s/'+i)

for i in os.listdir('val_images/moto_x'):
    shutil.copy('val_images/moto_x/'+i,'validation/Motorola-X/'+i)

for i in os.listdir('val_images/sony_nex7'):
    shutil.copy('val_images/sony_nex7/'+i,'validation/Sony-NEX-7/'+i)

for i in os.listdir('val_images/nexus_5x'):
    shutil.copy('val_images/nexus_5x/'+i,'validation/LG-Nexus-5x/'+i)

for i in os.listdir('val_images/samsung_note3'):
    shutil.copy('val_images/samsung_note3/'+i,'validation/Samsung-Galaxy-Note3/'+i)

for i in os.listdir('val_images/htc_m7'):
    shutil.copy('val_images/htc_m7/'+i,'validation/HTC-1-M7/'+i)

for i in os.listdir('./val_images/samsung_s4'):
    shutil.copy('./val_images/samsung_s4/'+i,'validation/Samsung-Galaxy-S4/'+i)

for i in os.listdir('./val_images/nexus_6'):
    shutil.copy('./val_images/nexus_6/'+i,'validation/Motorola-Nexus-6/'+i)

for i in os.listdir('./val_images/iphone_6'):
    shutil.copy('./val_images/iphone_6/'+i,'validation/iPhone-6/'+i)

for i in os.listdir('./val_images/moto_maxx'):
    shutil.copy('./val_images/moto_maxx/'+i,'validation/Motorola-Droid-Maxx/'+i)

for i in os.listdir('./flickr_images/iphone_4s'):
    os.rename('./flickr_images/iphone_4s/'+i,'train/iPhone-4s/'+i)

for i in os.listdir('./flickr_images/moto_x'):
    os.rename('./flickr_images/moto_x/'+i,'train/Motorola-X/'+i)

for i in os.listdir('./flickr_images/sony_nex7'):
    os.rename('./flickr_images/sony_nex7/'+i,'train/Sony-NEX-7/'+i)

for i in os.listdir('./flickr_images/nexus_5x'):
    os.rename('./flickr_images/nexus_5x/'+i,'train/LG-Nexus-5x/'+i)

for i in os.listdir('./flickr_images/samsung_note3'):
    os.rename('./flickr_images/samsung_note3/'+i,'train/Samsung-Galaxy-Note3/'+i)

for i in os.listdir('./flickr_images/htc_m7'):
    os.rename('./flickr_images/htc_m7/'+i,'train/HTC-1-M7/'+i)

for i in os.listdir('./flickr_images/samsung_s4'):
    os.rename('./flickr_images/samsung_s4/'+i,'train/Samsung-Galaxy-S4/'+i)

for i in os.listdir('./flickr_images/nexus_6'):
    os.rename('./flickr_images/nexus_6/'+i,'train/Motorola-Nexus-6/'+i)

for i in os.listdir('./flickr_images/iphone_6'):
    os.rename('./flickr_images/iphone_6/'+i,'train/iPhone-6/'+i)

for i in os.listdir('./flickr_images/moto_maxx'):
    os.rename('./flickr_images/moto_maxx/'+i,'train/Motorola-Droid-Maxx/'+i)
