==> 05_MobileNet_image_aug.ipynb <==
from keras.preprocessing.image import ImageDataGenerator
from keras.models import Sequential, Model
from keras.layers import Dropout, Flatten, Dense, Input
from keras import applications
from keras.layers import AveragePooling2D, GlobalAveragePooling2D
from keras.callbacks import ModelCheckpoint, ReduceLROnPlateau, EarlyStopping, CSVLogger, TensorBoard, LambdaCallback
from keras.applications.mobilenet import MobileNet, preprocess_input 
from keras.layers import Conv2D, Convolution2D, MaxPooling2D, ZeroPadding2D, BatchNormalization, Activation
from keras.optimizers import Adam
from keras import backend as K
import numpy as np
import pandas as pd
from keras import layers
import matplotlib.pyplot as plt
import matplotlib.image as mpimg
from keras.models import load_model
import random
import cv2
from PIL import Image
import os
from keras.utils import to_categorical
from collections import defaultdict
import jpeg4py as jpeg
from io import BytesIO

#parameters
classes = 10
batch_size=80
image_size = 512
train_total = 3500
validation_total = 875

model_mobilenet = MobileNet(include_top=False, weights = None,input_shape=(image_size,image_size,3))

x = model_mobilenet.output
x = GlobalAveragePooling2D()(x)
x = Dropout(0.5)(x)
x = Dense(256, activation='relu')(x)
x = Dropout(0.5)(x)
x = BatchNormalization()(x)
x = Dense(128, activation='relu')(x)
x = Dropout(0.5)(x)
x = BatchNormalization()(x)
x = Dense(classes, activation='softmax')(x)

for layer in model_mobilenet.layers:
    layer.trainable = True

model = Model(inputs=model_mobilenet.input, outputs=x)
model.summary()

train_datagen = ImageDataGenerator(horizontal_flip=True)
validation_datagen = ImageDataGenerator(preprocessing_function=preprocess_input)
                    

train_generator = train_datagen.flow_from_directory(
                    'train',
                    target_size=(image_size,image_size),
                    batch_size=batch_size,
                    class_mode='categorical',
                    shuffle=True
                    )


validation_generator = validation_datagen.flow_from_directory(
                        'validation',
                        target_size=(image_size,image_size),
                        batch_size=batch_size,
                        class_mode='categorical',
                        shuffle=False)

def image_aug_valid(image, index):
    Image.LOAD_TRUNCATED_IMAGES = True
    mpimg.LOAD_TRUNCATED_IMAGES = True
    if(index==0):
        res_image = np.uint8(cv2.pow(image/255.,0.8)*255.)
    elif(index==1):
        res_image = np.uint8(cv2.pow(image/255.,1.2)*255.)
    elif(index==2):
        res_image = cv2.resize(image,(0,0),fx=0.5,fy=0.5,interpolation = cv2.INTER_CUBIC)
    elif(index==3):
        res_image = cv2.resize(image,(0,0),fx=0.8,fy=0.8,interpolation = cv2.INTER_CUBIC)    
    elif(index==4):
        res_image = cv2.resize(image,(0,0),fx=1.5,fy=1.5,interpolation = cv2.INTER_CUBIC)
    elif(index==5):
        res_image = cv2.resize(image,(0,0),fx=2.0,fy=2.0,interpolation = cv2.INTER_CUBIC)
    elif(index==6):
        img = Image.fromarray(image)
        out_70_valid = BytesIO()
        img.save(out_70_valid, "JPEG", quality=70)
        res_image = jpeg.JPEG(np.frombuffer(out_70_valid.getvalue(), dtype=np.uint8)).decode()
        del img
        del out_70_valid
    elif(index==7):
        img = Image.fromarray(image)
        out_90_valid = BytesIO()
        img.save(out_90_valid, "JPEG", quality=90)
        res_image = jpeg.JPEG(np.frombuffer(out_90_valid.getvalue(), dtype=np.uint8)).decode()
        del img
        del out_90_valid
    return res_image 

def image_aug(image, index):
    Image.LOAD_TRUNCATED_IMAGES = True
    mpimg.LOAD_TRUNCATED_IMAGES = True
    if(index==0):
        res_image = np.uint8(cv2.pow(image/255.,0.8)*255.)
    elif(index==1):
        res_image = np.uint8(cv2.pow(image/255.,1.2)*255.)
    elif(index==2):
        res_image = cv2.resize(image,(0,0),fx=0.5,fy=0.5,interpolation = cv2.INTER_CUBIC)
    elif(index==3):
        res_image = cv2.resize(image,(0,0),fx=0.8,fy=0.8,interpolation = cv2.INTER_CUBIC)    
    elif(index==4):
        res_image = cv2.resize(image,(0,0),fx=1.5,fy=1.5,interpolation = cv2.INTER_CUBIC)
    elif(index==5):
        res_image = cv2.resize(image,(0,0),fx=2.0,fy=2.0,interpolation = cv2.INTER_CUBIC)
    elif(index==6):
        img = Image.fromarray(image)
        out_70 = BytesIO()
        img.save(out_70, "JPEG", quality=70)
        res_image = jpeg.JPEG(np.frombuffer(out_70.getvalue(), dtype=np.uint8)).decode()
        del img
        del out_70
    elif(index==7):
        img = Image.fromarray(image)
        out_90 = BytesIO()
        img.save(out_90, "JPEG", quality=90)
        res_image = jpeg.JPEG(np.frombuffer(out_90.getvalue(), dtype=np.uint8)).decode()
        del img
        del out_90
    return res_image 

def random_crop(image, crop_size):
    h,w,d = image.shape
    rand_num_h = random.randint(0,h-crop_size)
    rand_num_w = random.randint(0,w-crop_size)
    image_crop = image[rand_num_h:rand_num_h+crop_size,rand_num_w:rand_num_w+crop_size,:]
    return image_crop

def train_gen(train_files, train_classes, batch_size, target_size, imggen):
    min_batch_size = batch_size //10
    train_data = list(zip(train_files, train_classes))
    train_data_dict = defaultdict(list)
    for i in range(len(train_data)):
        train_data_dict[train_data[i][0].split('/')[0]].append(train_data[i])
       
    while(True):
        for i in train_data_dict.keys():
            random.shuffle(train_data_dict[i])
        for start in range(0, len(train_data), min_batch_size):
            image_crop_list = []
            image_classes_list = []
            
            
            for j in train_data_dict.keys():
                end = min(start + min_batch_size, len(train_data_dict[j])) 
                
                x_batch = [train_data_dict[j][i][0] for i in range(start,end)]
                
                for index,i in enumerate(x_batch):
                    image = jpeg.JPEG('train/' + i).decode()
                    
                    if(len(image.shape)<3):
                        print(_1_)
                        continue
                    rand_num = random.randint(0,7)
                    
                    augmented_image = image_aug(image,rand_num)
                    aug_image_recrop = random_crop(augmented_image,image_size)
                    if(np.random.rand()<0.1):
                        aug_image_recrop = np.rot90(aug_image_recrop,1,(0,1))
                    image_crop_list.append(aug_image_recrop)
                    image_classes_list.append(train_data_dict[j][start+index][1])
                    
                    del image
                    del aug_image_recrop
            
            if(len(image_classes_list)<batch_size-2*min_batch_size):
                break
            image_crop_list = [x/1. for x in image_crop_list]
            image_crop_list = [imggen.random_transform(x) for x in image_crop_list]
            image_crop_list = [preprocess_input(x) for x in image_crop_list]

            x_batch = np.array(image_crop_list, np.float32)
            y_batch = np.array(image_classes_list)         
            
            yield (x_batch, y_batch)

train_crop_generator = train_gen(train_generator.filenames,to_categorical(train_generator.classes),batch_size,image_size,train_datagen) 

def valid_gen(valid_files, valid_classes, batch_size, target_size, imggen):

    valid_data = list(zip(valid_files, valid_classes))
  
    while(True):
        
        for manip_num in range(8):
            for start in range(0, len(valid_data), batch_size):
                image_crop_list = []
                image_classes_list = []
                end = min(start + batch_size, len(valid_data)) 

                x_batch = [valid_data[i][0] for i in range(start,end)]

                for index,i in enumerate(x_batch):
                    if(manip_num == 2 or manip_num == 3):
                        temp_file = random.choice(os.listdir('validation_full_size/'+i.split('/')[0]))
                        image = jpeg.JPEG('validation_full_size/' + i.split('/')[0]+'/' + temp_file).decode()
                    else:
                        image = jpeg.JPEG('validation/' + i).decode() 
                    if(len(image.shape)<3):
                            continue
                    #rand_num = random.randint(0,7)

                    #augmented_image = image_aug_valid(image,manip_num)
                    aug_image_recrop = random_crop(image,image_size)
                    image_crop_list.append(aug_image_recrop)
                    image_classes_list.append(valid_data[start+index][1])

                    del image
                    del aug_image_recrop


                image_crop_list = [x/1. for x in image_crop_list]
                image_crop_list = [preprocess_input(x) for x in image_crop_list]

                x_batch = np.array(image_crop_list, np.float32)
                y_batch = np.array(image_classes_list)         

                yield (x_batch, y_batch)

valid_aug_generator = valid_gen(validation_generator.filenames,to_categorical(validation_generator.classes),batch_size,image_size,validation_datagen) 

model.compile(loss='categorical_crossentropy',
             optimizer=Adam(lr=2.5e-5),
             metrics=['accuracy'])

from pushbullet import Pushbullet
pb = Pushbullet('o.KiDDDXPuzV4qKbXh4Lywbgw1tK2oFfq1')

pushbullet_callback = LambdaCallback(
    on_epoch_end=lambda epoch, logs: pb.push_note("epoch: "+str(epoch),"train_loss: "+str(logs['loss'])+"    val_loss"+str(logs['val_loss'])))

callbacks = [ModelCheckpoint(filepath='mobilenet_image_aug_3.hdf5', verbose=1, save_best_only=True, save_weights_only=True),
ReduceLROnPlateau(monitor='val_loss', factor=0.5, patience=3, verbose=1,epsilon=0.01),
#EarlyStopping(monitor='val_loss', patience=10, verbose=1),
CSVLogger('./3-metrics_mobilenet_image_aug.csv'),
TensorBoard(log_dir='logs_mobilenet_image_aug', write_graph=True),
            pushbullet_callback]

history = model.fit_generator(
                    train_crop_generator,
                    steps_per_epoch = 20*int(np.ceil(train_total/batch_size)),
                    epochs=100,
                    validation_data=valid_aug_generator,
                    validation_steps= 8*int(np.ceil(validation_total/batch_size)),
                    verbose=1,
                    callbacks=callbacks)

model.load_weights('mobilenet_96_2_LB_unalt.hdf5')

model.evaluate_generator(valid_aug_generator,8*int(np.ceil(validation_total/batch_size)))

# validation predictions for analysis

predictions_valid = model.predict_generator(validation_generator,int(np.ceil(validation_total/batch_size)))
predictions_valid = np.argmax(predictions_valid, axis=1)

validation_generator.class_indices

true_positive = np.zeros(10)
false_positive = np.zeros(10)
true_negative = np.zeros(10)
false_negative = np.zeros(10)
for i in range(len(predictions_valid)):
    if(predictions_valid[i]==validation_generator.classes[i]):
        true_positive[predictions_valid[i]] = true_positive[predictions_valid[i]] + 1
    else:
        false_positive[predictions_valid[i]] = false_positive[predictions_valid[i]] + 1
        false_negative[validation_generator.classes[i]] = false_negative[validation_generator.classes[i]] +1

print('true_positive')
print(true_positive)
print('false_positive')
print(false_positive)
print('false_negative')
print(false_negative)
print('Total')
print(true_positive+false_negative)


#test predictions - 512

filenames_list = []
images_list = []
for i in os.listdir('test/temp/'):
    img = Image.open('test/temp/'+i)
    filenames_list.append(i)
    images_list.append(np.array(img, np.float32))
    

images_list = [np.rot90(x,3,(0,1)) for x in images_list]
images_list = [preprocess_input(x) for x in images_list]

new_array_arr = np.array(images_list)


new_array_arr.shape

test_preds_1 = model.predict(new_array_arr)

np.save('mobilenet_512_test_preds_1_unalt.npy',test_preds_1)

test_preds_2 = model.predict(new_array_arr)

np.save('mobilenet_512_test_preds_2_unalt.npy',test_preds_2)

test_preds_3 = model.predict(new_array_arr)

np.save('mobilenet_512_test_preds_3_unalt.npy',test_preds_3)

len(filenames_list)

test_preds_1_unalt = np.load('mobilenet_512_test_preds_1_unalt.npy')
test_preds_2_unalt = np.load('mobilenet_512_test_preds_2_unalt.npy')
test_preds_3_unalt = np.load('mobilenet_512_test_preds_3_unalt.npy')

test_preds_1_manip = np.load('mobilenet_512_test_preds_1_manip.npy')
test_preds_2_manip = np.load('mobilenet_512_test_preds_2_manip.npy')
test_preds_3_manip = np.load('mobilenet_512_test_preds_3_manip.npy')

for index,i in enumerate(filenames_list):
    if(i.split('_')[-1]=='manip.tif'):
        test_preds_1_unalt[index] = test_preds_1_manip[index]
        test_preds_2_unalt[index] = test_preds_2_manip[index]
        test_preds_3_unalt[index] = test_preds_3_manip[index]

test_preds_1_unalt.shape

np.save('mobilenet_512_test_preds_1.npy',test_preds_1_unalt)
np.save('mobilenet_512_test_preds_2.npy',test_preds_2_unalt)
np.save('mobilenet_512_test_preds_3.npy',test_preds_3_unalt)

predicted_classes_manip = [class_ids[np.argmax(i)] for i in test_preds]
predicted_classes_unalt = [class_ids[np.argmax(i)] for i in test_preds_unalt]

filenames_list_224 = []
images_list_224 = []
for i in os.listdir('test_224/temp/'):
    img = Image.open('test_224/temp/'+i)
    filenames_list_224.append(i)
    images_list_224.append(np.array(img, np.float32))
    
images_list_224 = [preprocess_input(x) for x in images_list_224]
images_list_224 = np.array(images_list_224)
images_list_224.shape

test_preds_224 = model.predict(images_list_224)

f = defaultdict(list)
for i,j in enumerate(filenames_list_224):
    f[j[2:]].append(i)

class_ids = {validation_generator.class_indices[x]: x for x in validation_generator.class_indices}

res_224_dict  = defaultdict(list)
for i in f.keys():
    x = np.mean(test_preds_224[f[i]],axis=0)
    res_224_dict[i].append(class_ids[np.argmax(x)])

result_filenames_224=[]
predicted_classes_224=[]
for i in f.keys():
    
    x = np.mean(test_preds_224[f[i]],axis=0)
    result_filenames_224.append(i)
    predicted_classes_224.append(class_ids[np.argmax(x)])

model.load_weights('resnet_image_aug_full.hdf5')

model.evaluate_generator(valid_aug_generator,8*int(np.ceil(validation_total/batch_size)))

test_preds_unalt = model.predict(new_array_arr)

test_preds_avg = (test_preds+test_preds_unalt)/2

result_filenames=[]
predicted_classes=[]
for i in f.keys():
    if(i.split('_')[-1]=='manip.tif'):
        x = np.mean(test_preds[f[i]],axis=0)
        result_filenames.append(i)
    else:
        x = np.mean(test_preds[f[i]],axis=0)
        result_filenames.append(i)
    predicted_classes.append(class_ids[np.argmax(x)])

predicted_classes = [class_ids[np.argmax(i)] for i in test_preds]

for index,i in enumerate(filenames_list):
    if(i.split('_')[-1]=='manip.tif'):
        predicted_classes_unalt[index]=predicted_classes_manip[index]

submission = pd.DataFrame({'fname':filenames_list,'camera':predicted_classes_unalt})
submission.to_csv('submission_mobilenet_512.csv', encoding="utf8", index=False)

from IPython.display import FileLink
FileLink('submission_mobilenet_512.csv')

# only mobilenet - unalt+ manip - Private LB score - 96.39
