==> 07_Ensemble_ml.ipynb <==
import numpy as np
import pandas as pd
from sklearn.cross_validation import train_test_split
from sklearn.model_selection import GridSearchCV
from sklearn.metrics import accuracy_score, make_scorer
from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import GridSearchCV
from sklearn.utils.class_weight import compute_class_weight
from keras.utils import to_categorical
from sklearn.linear_model import LogisticRegression
from sklearn.tree import DecisionTreeClassifier
from sklearn.svm import SVC
from xgboost.sklearn import XGBClassifier

# loading validations predictions for ensembling

filenames = np.load('filenames_list.npy')

resnet_preds = np.load('resnet_preds.npy')
mobilenet_preds = np.load('mobilenet_preds.npy')
cacnn_preds = np.load('cacnn_preds.npy')

resnet_preds_224 = np.load('resnet_preds_224.npy')
mobilenet_preds_224 = np.load('mobilenet_preds_224.npy')
cacnn_preds_224 = np.load('cacnn_preds_224.npy')

resnet_preds_classes = np.argmax(resnet_preds,axis=1)
mobilenet_preds_classes = np.argmax(mobilenet_preds,axis=1)
cacnn_preds_classes = np.argmax(cacnn_preds,axis=1)

resnet_preds_classes_224 = np.argmax(resnet_preds_224,axis=1)
mobilenet_preds_classes_224 = np.argmax(mobilenet_preds_224,axis=1)
cacnn_preds_classes_224 = np.argmax(cacnn_preds_224,axis=1)

mean_preds = (resnet_preds+mobilenet_preds+cacnn_preds+resnet_preds_224+mobilenet_preds_224+cacnn_preds_224)/6

np.bincount(np.argmax(mean_preds,axis=1)==target_classes)

feature_arr_classes = np.array([mobilenet_preds_classes,resnet_preds_classes,cacnn_preds_classes,mobilenet_preds_classes_224,resnet_preds_classes_224,cacnn_preds_classes_224])

max_voting = [np.argmax(np.bincount(feature_arr_classes[:,i])) for i in range(874)]

np.bincount(max_voting==target_classes)

cnt=0
for i in range(len(resnet_preds_classes)):
    if(mobilenet_preds_classes_224[i]!=mobilenet_preds_classes[i]):
        cnt = cnt +1

cacnn_preds_classes.shape

target_classes = np.load('target_classes.npy')

target_classes_one_hot = to_categorical(target_classes)

feature_arr = np.concatenate([mobilenet_preds,resnet_preds,cacnn_preds,mobilenet_preds_224,resnet_preds_224,cacnn_preds_224],axis=1)

np.bincount(mobilenet_preds_classes==target_classes)

features_df = pd.DataFrame(feature_arr)

# splitting the data into train test split

X_train, X_test, y_train, y_test = train_test_split(features_df, target_classes, test_size = 0.2, random_state = 0)

print(X_train.shape[0])
print(X_test.shape[0])

# ensembling with Random Forest

model_rf = RandomForestClassifier(n_estimators=25,verbose=0,oob_score=True)

parameters_rf = {'random_state':[10,20,30,40,50],'n_estimators':[2,4,7,10,20,40,80,100,150],'criterion':['gini','entropy'],'max_features':['auto','log2'],'class_weight':['balanced',None],'warm_start':['True','False']}

grid_obj_rf = GridSearchCV(estimator= model_rf, param_grid= parameters_rf, scoring= 'accuracy',refit=True,cv=3,n_jobs=6)

grid_obj_rf.fit(X_train, y_train)

grid_obj_rf.best_params_

accuracy_score(grid_obj_rf.best_estimator_.predict(X_test),y_test)

grid_obj_rf.best_score_

model_rf1 = RandomForestClassifier(n_estimators=35,verbose=0,oob_score=True,random_state=40)
model_rf1.fit(features_df,target_classes)
model_rf1.oob_score_

#logistic regression

model_lr = LogisticRegression(verbose=0,class_weight='balanced')

parameters_lr = {'random_state':[10,20,30,40,50],'C':[1.0,0.7,0.5,0.9],'solver':['newton-cg', 'sag', 'saga','lbfgs'],'multi_class':['ovr', 'multinomial'],'warm_start':[True,False]}

grid_obj_lr = GridSearchCV(estimator= model_lr, param_grid= parameters_lr, scoring= 'accuracy',refit=True,cv=3,n_jobs=6)

grid_obj_lr.fit(X_train, y_train)

grid_obj_lr.best_params_

accuracy_score(grid_obj_lr.best_estimator_.predict(X_test),y_test)

grid_obj_lr.best_score_

#decision tree

model_dt = DecisionTreeClassifier(class_weight='balanced')

parameters_dt = {'random_state':[10,20,30,40,50],'criterion':['gini','entropy'],'max_features':['auto','log2'],'min_samples_split':[2,5,7,10,13,15,20]}

grid_obj_dt = GridSearchCV(estimator= model_dt, param_grid= parameters_dt, scoring= 'accuracy',refit=True,cv=3,n_jobs=6)

grid_obj_dt.fit(X_train, y_train)

grid_obj_dt.best_params_

accuracy_score(grid_obj_dt.best_estimator_.predict(X_test),y_test)

grid_obj_dt.best_score_

#SVM

model_svm = SVC(class_weight='balanced',probability=True)

parameters_svm = {'random_state':[10,20,30,40,50],'C':[1.0,0.9,0.8,0.7,0.6,0.5,0.4],'kernel':['rbf','sigmoid','linear','poly']}

grid_obj_svm = GridSearchCV(estimator= model_svm, param_grid= parameters_svm, scoring= 'accuracy',refit=True,cv=3,n_jobs=6)

grid_obj_svm.fit(X_train, y_train)

grid_obj_svm.best_params_

accuracy_score(grid_obj_svm.best_estimator_.predict(X_test),y_test)

grid_obj_svm.best_score_

#xgboost

model_xgb = XGBClassifier(class_weight='balanced',objective='multi:softmax')

parameters_xgb = {'random_state':[10,20,30,40,50],'max_depth':[1,2,3,5],'learning_rate':[0.1,0.3,0.5],'n_estimators':[5,7,10,30,60,100,140],'booster':['gbtree', 'gblinear','dart']}

grid_obj_xgb = GridSearchCV(estimator= model_xgb, param_grid= parameters_xgb, scoring= 'accuracy',refit=True,cv=3,n_jobs=6)

grid_obj_xgb.fit(X_train, y_train)

grid_obj_xgb.best_params_

accuracy_score(grid_obj_xgb.best_estimator_.predict(X_test),y_test)

grid_obj_xgb.best_score_

#mean, max-voting

valid_preds_rf = grid_obj_rf.best_estimator_.predict_proba(features_df)
valid_preds_lr = grid_obj_lr.best_estimator_.predict_proba(features_df)
valid_preds_dt = grid_obj_dt.best_estimator_.predict_proba(features_df)
valid_preds_svm = grid_obj_svm.best_estimator_.predict_proba(features_df)
valid_preds_xgb = grid_obj_xgb.best_estimator_.predict_proba(features_df)

valid_preds_mean = (valid_preds_rf+valid_preds_lr+valid_preds_dt+valid_preds_svm+valid_preds_xgb)/5

np.bincount(np.argmax(valid_preds_mean,axis=1)==target_classes)

valid_preds_voting = np.array([np.argmax(valid_preds_rf,axis=1),np.argmax(valid_preds_lr,axis=1),np.argmax(valid_preds_dt,axis=1),np.argmax(valid_preds_svm,axis=1),np.argmax(valid_preds_xgb,axis=1)])

valid_preds_max_voting = [np.argmax(np.bincount(valid_preds_voting[:,i])) for i in range(874)]

np.bincount(valid_preds_max_voting==target_classes)

#ensembling with mean, and voting

mobilenet_test_preds = np.load('mobilenet_test_preds.npy')
resnet_test_preds = np.load('resnet_test_preds.npy')
cacnn_test_preds = np.load('cacnn_test_preds.npy')
mobilenet_test_preds_224 = np.load('mobilenet_test_preds_224.npy')
resnet_test_preds_224 = np.load('resnet_test_preds_224.npy')
cacnn_test_preds_224 = np.load('cacnn_test_preds_224.npy')

test_filenames = np.load('test_filenames.npy')

test_preds = np.concatenate([mobilenet_test_preds,resnet_test_preds,cacnn_test_preds,mobilenet_test_preds_224,resnet_test_preds_224,cacnn_test_preds_224],axis=1)

test_features_df = pd.DataFrame(test_preds)

test_preds_mean = []

for i in range(len(test_filenames)):
    test_preds_mean.append(np.mean(np.array([mobilenet_test_preds[i],resnet_test_preds[i],cacnn_test_preds[i],mobilenet_test_preds_224[i],resnet_test_preds_224[i],cacnn_test_preds_224[i]]),axis=0))

test_preds_rf = grid_obj_rf.best_estimator_.predict_proba(test_features_df)
test_preds_lr = grid_obj_lr.best_estimator_.predict_proba(test_features_df)
test_preds_dt = grid_obj_dt.best_estimator_.predict_proba(test_features_df)
test_preds_svm = grid_obj_svm.best_estimator_.predict_proba(test_features_df)
test_preds_xgb = grid_obj_xgb.best_estimator_.predict_proba(test_features_df)

test_preds_mean = (test_preds_rf+test_preds_lr+test_preds_dt+test_preds_svm+test_preds_xgb)/5

test_preds_voting = np.array([np.argmax(test_preds_rf,axis=1),np.argmax(test_preds_lr,axis=1),np.argmax(test_preds_dt,axis=1),np.argmax(test_preds_svm,axis=1),np.argmax(test_preds_xgb,axis=1)])

test_preds_max_voting = [np.argmax(np.bincount(test_preds_voting[:,i])) for i in range(len(test_filenames))]

final_test_preds = model_rf1.predict(test_preds)

class_map = {'HTC-1-M7': 0,
 'LG-Nexus-5x': 1,
 'Motorola-Droid-Maxx': 2,
 'Motorola-Nexus-6': 3,
 'Motorola-X': 4,
 'Samsung-Galaxy-Note3': 5,
 'Samsung-Galaxy-S4': 6,
 'Sony-NEX-7': 7,
 'iPhone-4s': 8,
 'iPhone-6': 9}

class_map = {class_map[i]:i for i in class_map.keys()}

class_map

final_test_preds = np.argmax(test_preds_mean,axis=1)

len(test_preds_max_voting)

sub_preds = [class_map[i] for i in final_test_preds]

test_filenames.shape

# submitting the final ensembled results

for index,i in enumerate(test_filenames):
    if(i.split('_')[-1]=='manip.tif'):
        sub_preds[index]=''

submission = pd.DataFrame({'fname':test_filenames,'camera':sub_preds})
submission.to_csv('submission_ensemble_mean_only_unalt.csv', encoding="utf8", index=False)

from IPython.display import FileLink
FileLink('submission_ensemble_mean_only_unalt.csv')

# Private LB score - 
# mean - 97.39
